AUT-2802: Collect PR metadata and add to s3 object
We now collect metadata from the PR that triggered the workflow and add
it to the S3 object metadata. This will allow us to use this data in the
CodePipeline execution to more easily corrolate a pipeline run with the
PR that triggered it.

This data is not technically required in the 'dev' workflow, but should
still be provided to ensure that we can keep the dev and prod workflows
consistent.

In future, we will be able to use this data to call back to GitHub and
add a comment to the PR / Commit with the URL of the associated pipeline
run.