AUT-2475: Remove request params from ticf stub
Unlike the request to the account interventions api, the ticf cri does not take a subject id in the request path. Instead all relevant information is submitted via the body, meaning that we don't need to define any terraform for this.

Additionally it should be noted (in case this changes in future) that this terraform had an issue and no mapping expression parameter was supplied for method.request.path.internalCommonSubjectId, so removing this also fixes an error when the terraform is run