AUT-2802: Prevent trigger change for dev deploy wf
This commit adds a GHA workflow to apply specific checks to GHA
workflows.

The first check is to ensure that the `deploy-api-modules-dev` workflow
is only triggered by a `workflow_dispatch` event. Devs may change this
trigger condition when working on a branch, but that change should not
be merged into the main branch.