BAU Refactor Audit Service.
Refactored the CheckReAuthUserHandlerTest to more efficiently create the expected AuditContext.

Extended the AuditContext to allow us to add new items to the AuditContext by calling a withXX method that creates a new record based on the existing record plus the new value.  We chose to do this rather than switch to using a POJO as having the AuditContext as a record still provides value.

Add the use of the AuditContext to the LoginHandler for the happy path and updated the unit test.

The LoginHandler is being refactored by creating an initial AuditContext and updating it alongside the existing AuditUser object.  When all the places in the code where the AuditUser is updated have shadow updates to the AuditContext we can remove the AuditUser and all references to it.