BAU: Remove redundant submitAuditEvent method
In a previous refactoring, we introduced a submitAuditEvent which took a txma audit user, as well as some other parameters. We've now gone for a slightly different approach, passing in an auditContext (which can contain values common to all requests in a handler but not coupled to what eventually ends up in the 'user' section of the txma audit event). So we can remove the unneeded submitAuditEvent method, and replace it with calls to our new method.

This leaves two submitAuditEvent methods. The eventual aim, once we have converted all handlers to use the new method, will be to remove the other one and leave a single interface for submitting an audit event. While we convert handlers one by one, we'll support the two ways of calling the audit service.