ATO-688: Add new DynamoDB table to store cached RP public keys
When an RP starts to share their signing key using a JWKS endpoint, Orchestration will need to retrieve and cache the key. This PRs adds a new DynamoDB table to store cached RP public keys. A key will be stored as a string in JWK format.

The table has a primary key schema with `clientId` as the partition key (HASH) and `keyId` as the sort key (RANGE).

The table is encrypted and TTL is enabled.

A new KMS key is defined to encrypt the table.