BAU: extract valid headers for mfa handler tests
This removes some noise from each test. In some cases, the test data being used was inconsistent for no good reason (for example, not setting a di persistent id, where this was not relevant to the test itself. This standardises the test data so there is less unimportant variance between tests