BAU: Remove duplicate test
This test was a hangover of when we had a state machine in the backend. However, now there is no code which is specific to uplift journeys in the MFA Handler. Following refactoring to remove common setup and noise in the mfa handler tests, sonarcloud is now flagging that this test is a duplicate of 'shouldReturn204ForSuccessfulMfaRequestWhenNonResendCode'. We can therefore remove this test