AUT-2851: Move TICF CRI Request file to shared
This will be used both as the input to the TICF CRI Stub, and as the request object sent by the TICF CRI Handler. Moving to a shared place means that these two bits of code are working from a shared schema, and that changes can be more easily coordinated.