AUT-2908: Add optional authenticated parameter to account interventions request
This will support the call to the ticf cri (which for now we will do from the account interventions handler, only in the case where the user is authenticated).

This field is currently unused - later work will implement logic based on this, so this is just to ensure that the frontend is making a compatible call to the backend.