BAU: Convert the Verify Code Handler to use the new submitAuditEvent
This cleans up the call sites by introducing a common audit context to be shared by all audit events submitted by the handler.

This also means we can remove an existing helper method to submit an audit event that was specific to this handler, which essentially solved the same problem as our new method.