AUT-2775: Correct method to get userId for PendingEmailCheckRequest
When on the `ACCOUNT_MANAGEMENT` journey to update the user email, we have a user profile and should add the id to `PendingEmailCheckRequest`.

The current way of finding the userId in `SendOtpNotificationHandler` looks up the profile form the email address we are sending the OTP too.

This has an issue because the user does not have that email yet. Instead, this commit looks up the userId from the authorizer principleId.

This is used in the same file to tie the SEND_OTP audit event to the same user.