BAU: Small refactor to handle notification request
We can retrieve the session from the user context which we're already passing through here.

This makes sonarcloud happy as it wants a maximum of seven parameters, and the audit context was the eighth.

In general, the send notification handler is probably ripe for a refactoring since these methods are quite chunky, but for now this just fixes the parameter problem.