BAU: Introduce constructor method for audit context
In most handlers, half of the fields in the audit context come from the user context. Therefore, creating a constructor which handles the retrieval of values from the user context makes the handler code less verbose.

I used my editor to optimise the imports in some handlers, which also reversed some wildcard imports, which we should be avoiding.