AUT-2495: Temporarily bump the ticf http timeout in staging
Bumping this to a large number (20 seconds) while we test out the integration between us and ticf in staging, as at the moment we're hitting timeouts from the api. The cold start for the ticf lambda is currently 13 seconds, which is being worked on.

The aim is to get the call to ticf working, and then reduce the timeout down to the agreed 2 seconds