BAU: Add dependency-related workflows
Add two new workflows to the repository that will help us manage
dependencies better. We now create and push a dependency graph to the
repository on push to main, which will allow better tracking of
dependencies. We also have a workflow that will run on pull requests
that will compare the new graph against `main` and warn if any
vulnerable dependencies / versions are being introduced.