BAU: Increase visibility timeout for pending_email_check_queue
During deployments of `authentication-check-experian` we sometimes see warnings when deploying a SQS Lambda trigger from this queue. The error warns that the lambda timeout is more than the visibility timeout of the queue.

The visibility timeout should be more than the lambda timeout, to prevent a second lambda being triggered by the same message while the first is still processing it.

This increases the queue visibility timeout past the lambda timeout of 45 seconds.