AUT-3373: Fixed incorrect DELETE_ACCOUNT audit context
Before this commit we were not correctly sending the persistent session ID in the audit context.
This is because persistentSessionId was never being reassigned after initially being assigned to null,
instead incorrectly calling the extractPersistentIdFromHeaders and assigning the result to ipAddress.

This fixes that and adds makes sure that they are tested in the audit context