BAU: Add SonarCloud break glass mechanism
Now that a successful SonarCloud analysis run is required to merge a PR, we have discovered a need to have a "break glass" mechanism to bypass the analysis in cases where we must push code quickly without an analysis on code smells / code coverage.

This means we can merge PRs that revert previously merged work that would now by hit by the failed analysis.

This is implemented by allowing the GitHub `run-sonar-analysis` job to be skipped if the PR has the label `bypass-sonarcloud`.

Various ways to do this were explored and discounted:

- Using the `if` key within a job would mean the job is skipped if true, which would fail the branch protection rule that requires a pass
- Trying to exit a job early didn't work either:
  - Having a step with a `run` key that used `exit 0` didn't work, as the step completed and the next step was started.
    - We could make it exit with an error code, but that would fail the job, thus fail the branch protection rule.
  - Using the `gh` CLI tool in `run` block to cancel the job failed, as it didn't have the right permission, also the job needs to succeed for the branch protection rule.
- Instead each step needs to check if it can be run with the new job environment variable.
  - Here we elevate the `github.actor != 'dependabot[bot]` condition that was in the last step into all steps, as the intent behind the condition is the same as the new label (skip the check but mark as soccess)