BAU: Retrigger pre-merge-checks when the pull request labels change
The previous commit adds a way to bypass the SonarCloud analysis if the PR has a label. Instead of requiring the PR author to jump through some hoops to get jobs to rerun after they assess the need to bypass this will now trigger the checks when the label is added/removed.