BAU: Remove test references to session storage count
These have long been redundant, as we've switched over to using the dedicated code storage redis store for getting counts. So for the places where we set this up in test, it wasn't required, and for the places where we used it as an assertion, this assertion is not actually testing anything - the relevant thing to test is either that the count done via the storage service is a particular number, or that we've deleted the count via the storage service. All these assertions are already in place.