AUT-2789: Log user out when they enter their password incorrectly max allowed times.
This is the backend piece of this feature.  Rather than setting a temporary lock against the users email when they fail to enter a valid password during the re-authentication flow the count of failed attempts will be cleared and the frontend nortified that the user has failed too many times.  The expected behaviour in the frontend is to then log the user off.  If the new re-authentication journey is not supported then the user will be blocked as before and the frontend expected to display the locked out page to the user.

Addressing Sonar issues and PR comments to simplify main handler method by introducing another clarifying private method call and removed a duplicate constant declaration.