BAU: Move pr-data to reusable workflow
Rather than duplicating the same step in both prod and dev workflows,
move it to a reusable workflow instead.

Also, use the short commit sha in the summary, to reduce the length of
the string in AWS - it is annoying when it's truncated. For the same
reason, swap the order of PR number and repository name - so we
front-load as much 'useful' stuff as possible.