BAU: Add detect-secrets pre-commit hook
This will scan for potential secrets inadvertently added to code.

`.secrets.baseline` contains an initial baseline of all existing
'violations' in the repository, so we don't immediately have hook
failures. These are all false positives.

A lot of these false positives are coming from the java tests. To avoid
creating new detections in future, there are a few options:

1. Move all the test passwords to a single file in `shared` and import
   them in the tests. Using these variables will stop the detections.
2. Add the 'ignore secret' comments when something is detected:
https://github.com/Yelp/detect-secrets/tree/master?tab=readme-ov-file#inline-allowlisting