BAU: Add cloudwatch metric for notify request duration
This will help us have visibility as to whether there are any slow downs in notify service, as this isn't immediately obvious from the notify dashboard, and will help us have visibility as to whether problems are on our side or notify's.

Errors parsing dates will not throw errors in the delivery receipts lambda for now, since this isn't critical information for the moment.

We only send this metric in the case of a successful delivery, since in the case of unsuccessful deliveries we might expect them to take longer to resolve due to retries. If there is a widespread issue with lots of unsuccessful deliveries, we already track these metrics via the existing dashboard.