ATO-952: Parse post logout url as part of validation
An invalid URL should be handled like any other input error.

It would have likely been sufficient to keep the url empty on the request object in the case the provided value did not match the client registry, but the chosen approach also protects against bad values in the client registry.