AUT-3046: Reauth - Remove user lockout from requesting too many sms code.
If the user is on a reauth journey and reauth is supported then they will not be locked out if they request too many security codes.

Clarify logic for decision to block user or log them out.  This uses the reauthentication supported feature flag which is mapped in the terraform configuration.