AUT-3481: Implement basic functionality of new addMfaMethod method
This method will eventually allow us to store multiple mfa methods. For now, it just inserts either an auth app or an sms method in the same way as the existing code.

Eventually the idea is that this replaces the existing updateMfa method as the canonical way of updating mfa