AUT-3481: Allow adding multiple MFA methods with different priorities
An MFA method will be appended to existing methods as long as it has a different priority to the existing method - if not it'll be overwritten.

Note that at the database level, we do allow the following which are not technically valid:
* a backup and default auth app (at a domain level, you cannot have two auth app methods)
* a backup and default sms method with the same phone number (again this would be invalid at the domain level)

This is because our current thinking is that we want to keep the database layer relatively straightforward, and implement further business rules at a different level (e.g. in a handler).