BAU: Rework pre-commit
Spotless now has a `ratchetFrom` config setting. This
allows you to specify a gitref to run spotless from (ie. it'll only
check files that have changed since the ref). I've set this up to pull
a ref value from a System Property (`ratchetFrom`), so it can be
switched on when required. Pre-commit passes `main`, to replicate the
normal pre-commit behaviour of only doing things with changed files.

Additionally:
- Replace the python linting suite with `ruff` - it comes
  'batteries included' and is faster than the previous set of tools
- Stop using the govuk-one-login pre-commit hooks: they use docker
  under the hood, which makes things super slow.