AUT-3481: Make Mfa Methods generic
We will be storing both AuthAppMfa and SmsMfa as mfa methods in the user credentials table.

Currently, we only store AuthAppMfa methods. This commit is an initial step, which should have no behavioural impact on anything, to make mfa methods generic.

This will allow us to store and work with types that are specific to the data being stored - for example, auth app mfa data has a credential value, whereas sms doesn't. Being able to work with different types means less annoying null handling.

We had to define a custom converter for the serialisation and deserialisation from the dynamo db objects, since we're now deserialising an interface which may take multiple forms