AUT-3481: Rename MFAMethod AuthAppMFAMethod
This better reflects the fact that this is currently only used to store auth app mfa.

It also serves as a step towards implementing other types of mfa, like sms. The idea is to introduce an MFAMethod interface implemented by both this Auth App Mfa and the new SMS Mfa, meaning that a user can have either type, but we can also deal with the correct object at the type level without having to handle nulls on fields which are only relevant to one.