AUT-3481: Make user credentials return generic MFA Method type
This will allow us to introduce different mfa method types stored in the same place but with their own specific types to work with at code level.

This necessitated introducing a custom converter for dynamo db to serialise and deserialise items.

This also means we have to do some casting when we read items out of the database.

Currently, we only support auth app mfa methods, but we'll be introducing sms mfa methods subsequently