AUT-3481: Make mfa identifier nullable
Makes the integer a boxed integer. This has the effect of not setting the value in the resulting data stored in the database - without this, it defaults to empty integer i.e. 0. Ensuring this is just not set when not explicitly set means that we're not making any choices now about what 0 means as an identifier and ensures that we don't have any nasty surprises down the road when we store an mfa method via the new method with an explicit value of zero.