AUT-3434: Fixing dependabot alerts raised by uk.gov.service.notify:notifications-java-client by upgrading to the latest version.
Added constraints to the dependency block.  To apply the constraints the java plugin needs to be in scope within the subprojects block.

Moved all the required versions to constraints.  Constraints do not pin us to a specific version of a library they define the minimum version that can be used.  Gradle's conflict resolution rules will use the latest version available from all transitive dependencies over-riding the constraint.

Dependabot issues:

[7](https://github.com/govuk-one-login/authentication-api/security/dependabot/7)
[10](https://github.com/govuk-one-login/authentication-api/security/dependabot/10)
[17](https://github.com/govuk-one-login/authentication-api/security/dependabot/17) jose4j vulnerabilities
[13](https://github.com/govuk-one-login/authentication-api/security/dependabot/13) Java: DoS Vulnerability in JSON-JAVA