AUT-3433: Fixing dependabot alerts raised by au.com.dius.pact.provider:junit5.
Added constraints to the dependency block.  To apply the constraints the java plugin needs to be in scope within the subprojects block.

Moved all the required versions to constraints.  Constraints do not pin us to a specific version of a library they define the minimum version that can be used.  Gradle's conflict resolution rules will use the latest version available from all transitive dependencies over-riding the constraint.

Links to dependabot issues:

[14](https://github.com/govuk-one-login/authentication-api/security/dependabot/14) Apache Commons IO vulnerabilities
[11](https://github.com/govuk-one-login/authentication-api/security/dependabot/11) [12](https://github.com/govuk-one-login/authentication-api/security/dependabot/12) Guava vulnerabilities
[9](https://github.com/govuk-one-login/authentication-api/security/dependabot/9) [18](https://github.com/govuk-one-login/authentication-api/security/dependabot/18) netty vulnerabilities
[15](https://github.com/govuk-one-login/authentication-api/security/dependabot/15) [16](https://github.com/govuk-one-login/authentication-api/security/dependabot/16) Apache Commons Compress vulnerabilities