AUT-3523: Change units of time of initial request from seconds to milliseconds
Eventually, this is used to increment a cloudwatch metric at the end of a request that records how long the request took. This metric was incorrect, as when we send the request we treat it as epoch seconds, but when we read it we treat it as epoch milliseconds, resulting in a nonsensical duration.

We could have fixed on either side, but if we set this as seconds we may lose some granularity, so I've gone with the approach of setting everything to milliseconds.