AUT-2418: Create endpoint-module-v2
This new endpoint module is a drop-in replacement for the existing
`endpoint-module`. It delegates all the lambda logic to the
`endpoint-lambda` module, while still taking api-gateway attributes
as variables.

This is here for two reasons:

1. The `endpoint-lambda` will be the future standard for creating
lambdas. Using this submodule allows us to create the same resources
that will be used after migration to the openapi gateway

2. Instead of updating the `endpoint-module` that all other API
components use, we keep the changes in this ticket exclusively to the
`account-management` component. It would work for them too, but it
seems pragmatic to only migrate one thing at a time.