AUT-2418: Initial stage of openapi migration
TL;DR:
- Create a second API Gateway, provisioned from an OpenAPI spec.
- Provision account-management lambdas with new `endpoint-module-v2`.

We provision a second API Gateway for Account Management. This new
gateway is provisioned via the `openapi-api-gateway` module.

The lambdas are now provisioned with the new `endpoint-module-v2`. This
allows us to retrieve the integration URIs to be used in the OpenAPI
spec.

The .tfvars specify that the new gateway will be dormant: The domain
will still be pointed at the existing API Gateway for now.

All localstack references have been removed.

Two `lambda-provisioning` files were renamed, to ensure that the
filenames match the endpoint name.