AUT-2418: Initial stage of openapi migration
TL;DR:
- Create a second API Gateway, provisioned from an OpenAPI spec.
- Provision account-management lambdas with new `endpoint-module-v2`.

We provision a private API Gateway for Account Management. This new
gateway is provisioned via the `private-api-gateway` module. This
gateway does not have an associated domain, but is accessed via the
VPC endpoint.

The lambdas are now provisioned with the new `endpoint-module-v2`. This
allows us to retrieve the integration URIs to be used in the OpenAPI
spec.

The .tfvars explicitely configure the OpenAPI spec file to use.

All localstack references have been removed.

Two `lambda-provisioning` files were renamed, to ensure that the
filenames match the endpoint name.

Two checkov rules have been marked for global skip here:
- CKV_OPENAPI_4: AWS deals with filling out the security bits here, as
  we are using an authorizer.
- CKV_OPENAPI_20: Not relevant for us. It's always going to be 'https',
  and we don't specify URLs in the OpenAPI spec.