BAU: Bump notify client from 3.19 to 5.2
This required us to add apache commons lang as an explicit dependency. It was previously being transitively pulled in by the previous version of the notify client. When we bumped the notify client, we no longer had this and the code that was using it failed to compile

From a review of the changelog, this bump of two major versions contains no other breaking changes for our use cases.