wip: sample method for method management
This is here to demonstrate how to add a new endpoint to the method
management API gateway.

The sample handler is essentially an echo. It can handle multiple
request types, with and without path parameters. This may be
unusual, but it's at least possible now! No more needing to create
a separate handler for every HTTP method!