BAU: Fix failing reauth test.
This was a new test that didn't get updated to reflect the new name of the audit event it emits.  It made it onto main and was deployed to prod because failing tests weren't causing a drop-out for the checks on main.  We have added an additional check to the GH workflow to address this.