AUT-3599: Retain reuath counts when a user exceed maximum retries
For reauth, we are going to use the counts to determine if someone is blocked, rather than separately maintaining a block.

Removing this now, since we're going to start returning block information for reauth in the start handler, and this will not be available to us if we delete the counts here.