AUT-3529: Added SMS code reauth attempt count
Implemented the create or increment count logic into the verify code handler, incrementing the count for
ENTER_SMS_CODE in the authentication attempts table for a reauth journey when a journey with NotificationType.MFA_SMS
fails. If there is a count for a user having failed to enter the right sms mfa code and they then succeed, the count
will be deleted.

Alongside this, this commit set environment variables in verify_code.tf and authdev and staging environments.