AUT-3613: Return different error response for password retries exceeded in reauth
In the new world, we want to log someone out rather than lock someone out when they've exceeded the max retries for reauth credential entry. Returning the existing code will mean that we treat this as a log out on the frontend. I've repurposed the error description with the view that we can return this error code for all reauth incorrect credential entries and treat these as non lock outs.

Also got rid of unnecessary mock checks in unit test