AUT-3613: Increment count of failed emails before throwing any errors
We want users to be logged out on the 6th attempt at an incorrect email in reauth. This is currently happening, but the count in the database is stopping at 5, which means that eventually the data we put into audit events will be confusing.

In this change, we increment the count of failed emails regardless of whether the current incorrect attempt logs someone out or not. We also set the configured value of max retries to 6, rather than 5. Together, these changes should not change the behaviour observed (a user should still be logged out on the 6th incorrect attempt), but should:

a) make the logic more consistent - currently max email attempts is configured to 5, but max passwords is configured to 6, but in both cases we only get locked out on th 6th incorrect attempt
b) make sure the value in the database is correct