AUT-3530: Introduce helper method to filter reauth counts by counts exceeded
Together with the new method introduced on the authentication attempts service, this gives a more flexible way to check whether counts are exceeded. Subsequently we can switch all the existing usages of 'isBlockedForReauth' over to this method, remove the existing method and make the helper look like more of a helper (static methods, not side effects).

The idea of returning all the counts leaves it flexible to the caller as to whether they want to do anything special with a certain count exceeding. For example, at the point of entering a password, they might want to respond differently in the case of an enter password count exceeding