ATO-981: Logs session Id if present in headers
We are now passing the sessionId from orch to auth, in the headers
of the http request to this userinfo endpoint. However these won't
line up perfectly, so we're logging the sessionId conditionally
if it is attached in the headers of the request. Eventually all
request will include this value, and might be able to attach this
value to the logs.