ATO-981: Retrive session from session store
As we now have the sessionId, we can now retrieve the User's session from
the shared session store. This is needed so we can eventually update the
isNewAccount value to EXSISTING at the end of the userinfo request. This
also updates the userInfo integration test to pass in the session id header
and reflect our new behaviours:

i) Returning a 400 when there is no session id in the headers

ii) Returning a 400 when no session is found for the sessionId

iii) Sets up the  session with a session helper function which
the account state to simulate the a new account journey.
This sets us up to be able to set the account status to existing
in subequent changes.