ATO-981: Sets isNewAccount to EXISTING
We are now setting the value of isNewAccount in the user's session
to EXSISTING. This will only affect new accounts as exsisting accounts
should have the value set to EXSISTING  in the Login Handler. Previously we relied
on orchestration setting this value in a shared session store, however
we're moving to our own session store, which orchestration cannot update
once the user is handed back to them.

This value will be set by Auth without relying on orchestration to set it,
making the subsquent updates by orchestration redundant. This change is needed
to ensure that if a user comes back within the same session, their account is now
marked as Exsisting in their session, not NEW. This is preparatory work to making the
sessions independent, as once auth set this value independently, we start to transition
to the new session store.