BAU: Introduce constant for request event
Extracts a constant to remove some noise in tests. Also, removes a bit of confusion, as we were duplicating some of the request information in both the api request event, and also the CheckReauthUserRequest. However, it's the data in the CheckReauthUserRequest that determines the behaviour in the handler when using handleRequestWithUserContext - this is the method that utilises the parsed request object that's been parsed from the event body in the base handler