AUT-3602: Add reauth_failed audit event in login handler
If the authentication attempts service is turned on, the reauth_failed audit event will
send in the event that a user has failed to reauthenticate successfully and has then been
locked out of their account. A subsequent commit will add a reauth_failed audit event
to be emitted in the case that the user is already been locked out of their account
before trying to reauthenticate.