AUT-3743: Add reauth counts to session before deleting from code store in Verify Code Handler
At the point of successful mfa entry in the mfa code handler, we want to delete the reauth counts from the code store, which controls blocking behaviour. However, we need to preserve the counts for audit purposes until the auth code handler, where we can say that a reauthentication was successful and emit the REAUTH_SUCCESS audit event.

This allows us to temporarily store these counts for audit purposes, without affecting the log out behaviour. We'll delete in the auth code handler.

Note here that we're duplicating retrieving the counts again. I've done this for now while other changes are being made to the codebase, but once things have settled we can do a final refactor to retrieve counts once per handler