ATO-981: Adds updateSession method
Adds an update session method and also adds some tests which
assert that the updated session object is saved. This adds some unit
tests alongside the integration tests. To make this possible I've added
a new constructor overload to the Auth session service class.

Also adds a test which asserts that when a previous session is updated with a
new sessionId, this still has the same other fields. We couldn't do
this before as the only field on the session was sessionId.