BAU: Refactor start handler and service
Refactored the start handler and service so that the counts from the authentication attempts service are only queried
once from the handler. Previously, getting the counts took place in the service so that a isBlockedForReauth could
be calculated and returned in the UserStartInfo. Since then we have added the AUTH_REAUTH_FAILED audit event
which requires the counts. It made more sense to call the authenticationAttemptsService.getCountsByJourney() method
to get the counts and calculate a value for isBlockedForReauth in the handler to be passed through to the service,
rather than pass back the counts from the service to the handler OR call the service twice.

Alongside this change has been a general shuffle of assignments in the handler and amending of tests to reflect the new flow.