BAU: Add timestamps to bookend email check logs
Adding timestamps to our existing logs for the start and end of the async email check process to help diagnose a potential issue on the `EmailCheckDuration` metric.

We're currently seeing 10 unaccounted for seconds in the async journey. After some analysis we're leaning to an issue with the `EmailCheckDuration` metric rather than those 10 seconds actually existing.

Adding these will allow a comparison with the value of the start and end times as recorded in the metric vs when the time the logs are recorded.