ATO-1064: moves non-docapp logic out of docapp flow
Instead of doing browser session id checks before the docapp journey continues, we instead move the docapp return before it to early return. We don't care about the bsid cookie for docapp, so this means we only log and audit bsid things for auth journeys, as well as separating concerns.