ATO-981: Set isNewAccount in userinfo handler
We can now set the value of isNewAccount in the new session store in
the userinfo handler. This completes our co-setting alongside the shared
session store.

Adding a new test for the session setting means we need to clear the
integration test audit queue between each test run to ensure we get
accurate counts of messages on the queue.