ATO-979: Populate userinfo response with claims from session
Originally we were just populating userinfo using data from the access token. Now we are also adding claims using data from the session, which Orchestration need.

For now we are only setting verified_mfa_method_type, but eventually any session data that needs to be sent to Orchestration will be included in addClaimsFromSession.