ATO-1067: Use message hash when signing long messages
Because we're using KMS to sign the JAR we send to Auth's /authorize
endpoint, there is a limit of 4096 bytes[1] as to how long the JAR
can be. Because we include RP set fields such as state and nonce,
we cannot predictably say how long the total length will be, and
this can cause us to run into KMS errors.

Instead here we introduce a check to see if the encoded message is
more than 4096 bytes, and if so we use a hash digest in
the sign request instead of a raw sign request.

[1] - https://docs.aws.amazon.com/kms/latest/APIReference/API_Sign.html#:~:text=To%20sign%20a%20larger%20message%2C%20generate%20a%20hash%20digest%20of%20the%20message%2C%20and%20then%20provide%20the%20hash%20digest%20in%20the%20Message%20parameter.%20To%20indicate%20whether%20the%20message%20is%20a%20full%20message%20or%20a%20digest%2C%20use%20the%20MessageType%20parameter