ATO-981: Remove isNewAccount setting from userinfo handler
Auth do not  read the value from the shared session store so, we should
stop setting it.

I have added a comment which reminds us to come back and remove reading
the redis session in this handler. I am leaving the reading of the redis
session in the code for now as we may need to set some values in the shared
session in the interim, as we transition fields over to the new session store.
Once we have completely removed this, we can remove reading redis from here.