AUT-3578: Check reauth rp pairwise id counts when subject id does not exist
When there is no subject id, it means the user does not have a session (they are not logged in). However, even for a non-logged in user we can still check whether they are blocked for reauthentication based only on the counts against their rp pairwise id, since attempts to enter a reauth email will be stored against this identifier for a non logged in user. So we can still determine whether the user might be blocked for reauth based on these counts (at a later point in the reauth journey we can check against both subject id and rp pairwise id counts).