AUT-3751: The MfaHandler has a guard to check when it is invoked by a User that is locked out due to presenting too many incorrect security codes.  This guard has been renamed to better reflect what it is achieving.
Once the count of security code requests has been incremented for the current request the count of times a security code has been requested must be checked to see if the User has exceeded the maximum allowed.  If the User has exceeded the maximum number of requests then we should return to the front end without sending a new security code.