AUT-2257: Fix Index Not Found exception on getUserProfileFromSubject and public subject methods
In this query, we were getting a stream of pages of dynamo db results, finding the optional first of these pages, and then assuming this would always have a first result. This is largely true (if there are no items, then normally duynamo db returns no pages). However, there seem to be some edge cases where we can hit this, and very occasionally (about once a month) we are hitting this error in production).

This takes the approach of the below function by treating both the first page, and the first item within that page, as optional, which should fix these errors.

Since this only happens in rare edge cases, it's hard to write a test to force out this change.

Some of the edge cases mentioned when researching this that would produce this result (not certain which of these caused our errors):

* Querying for a non existent key in a sparse secondary index
* A bulk delete
* some other cases that I don't think can have applied here, but with the general lesson being that we can't guarantee no pages will be returned when there are no matching results